// src/lib/sanity/sanity.types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type PagesQueryResult = Array<{
  _id: string;
  title: string | null;
  slug: string | null;
  excerpt: string | null;
  _updatedAt: string;
}>;
// ...
export type PostsQueryResult = Array<{
  _id: string;
  title: string | null;
  slug: string | null;
  excerpt: string | null;
  publishedAt: string | null;
  _updatedAt: string;
  image: {
    _type: "imageWithAlt";
    alt: string | null;
    caption: string | null;
    hotspot: SanityImageHotspot | null;
    crop: SanityImageCrop | null;
    asset: {
      _id: string;
      url: string | null;
      metadata: {
        dimensions: {
          width: number | null;
          height: number | null;
        } | null;
        lqip: string | null;
      } | null;
    } | null;
  } | null;
}>;